RavenDB-4802 - don't set cache when using precomputed batch
